fix: quickstart fixes

+++ scripts/quickstart.ts
@@ -1,6 +1,12 @@
 import { $ } from "bun";
 import * as Bun from "bun";
 import {
+	getVersion,
+	isCommandAvailable,
+	SetupASDFPaths,
+	versionSatisfies,
+} from "./utils";
+import {
 	bgDarkGray,
 	yellow,
 	red,
@@ -9,32 +15,61 @@ import {
 	darkGray,
 	magenta,
 } from "ansicolor";
+import packageJson from "../package.json" assert { type: "json" };
 
 console.log(`${lightGray("🦨💕 Quickstart Installer...")}`);
 
-if (process.platform === "darwin") {
+if (!versionSatisfies(await $`bun --version`.text(), packageJson.engines.bun)) {
+	prompt(`Press enter to upgrade ${bgDarkGray(" 🍞 bun ")}`);
+	await $`bun upgrade`;
+	console.log(`${lightGray("✅ Bun upgraded, please restart this script")}`);
+	process.exit(0);
+}
+
+// check if asdf is installed using which asdf
+if (!(await isCommandAvailable("asdf"))) {
 	prompt(`Press enter to install ${bgDarkGray(" asdf ")} using HomeBrew`);
 	await $`brew install asdf`;
 }
 
-const versionFile = (
-	await Bun.file("./packages/contracts/.tool-versions").text()
-).split("\n");
+if (
+	!versionSatisfies(await $`asdf --version`.text(), packageJson.engines.asdf)
+) {
+	prompt(`Press enter to update ${bgDarkGray(" asdf ")} using HomeBrew`);
+	await $`brew upgrade asdf`;
+}
 
+if (!(await Bun.file("./.tool-versions").exists())) {
+	throw new Error("No .tool-versions file found");
+}
+
+// read asdf tool versions
+const toolVersions = Object.fromEntries(
+	(await Bun.file("./.tool-versions").text())
+		.trim()
+		.split("\n")
+		.map((x) => {
+			const v = getVersion(x.trim());
+			const [key] = x.split(" ");
+			return [key.trim(), v];
+		}),
+);
 prompt(
 	`\nPress enter to install ${yellow(" 🐞 scarb ")} and ${red(" ⛩️ dojo ")}`,
 );
 
-await $`asdf install scarb ${versionFile[0].replace("scarb ", "")}`;
-await $`asdf install dojo ${versionFile[1].replace("dojo ", "")}`;
-
+console.log(`asdf install scarb ${toolVersions.scarb}`);
+await $`asdf install scarb ${toolVersions.scarb}`;
+await $`asdf set scarb ${toolVersions.scarb}`;
+await $`asdf install dojo ${toolVersions.dojo}`;
+await $`asdf set dojo ${toolVersions.dojo}`;
 await $`asdf current`;
-
+await $`asdf reshim dojo`;
+await $`asdf reshim scarb`;
+// await SetupASDFPaths();
 console.log(`${lightGray("✅ Done!")}`);
 
-import pck from "../package.json" assert { type: "json" };
-
-const runScripts = Object.entries(pck.scripts).map(([key, value]) => {
+const runScripts = Object.entries(packageJson.scripts).map(([key, value]) => {
 	const name = value.split("#")[0].trim();
 	const description = value.split("#")[1].trim() || "";
 	return {
